refactor(core): ms时间戳字段 u64→u32 — bindings 输出 clean number 而非 bigint
将 ms 时间戳家族的存储类型从 u64 收窄至 u32（值域 49.7 天, 足够所有视频时长场景），
bindings 从 bigint 统一变为 number, TS 侧不再需要 BigInt 处理。

变更范围（跨 6 阶段 + 服务器）:
- asr: AsrWord.start/end, AsrSegment.start_ms/end_ms, AsrResultMeta.audio_duration,
       WhisperOffset.from/to → u32
- translate: TranslateSegment/TranslatePartialSegment start_ms/end_ms → u32
- split_audio: SplitAudioTiming.start_ms/end_ms, SplitAudioSegment.split_start_ms/split_end_ms,
               SplitAudioArgs.start_pad_ms/end_pad_ms → u32
- tts: TtsSegment.slot_end_ms/tts_duration_ms → u32
- mix_audio: Timing 中所有 ms 字段 → u32, drift_ms → i32
- servers: ModelServerStatus.uptime_s → u32
- utils: SrtSeg 及 srt_time/parse_srt_time → u32

联动修复:
- asr/mod.rs: word timestamp offset 计算结果 as u32, segment_rms 参数改为 u32
- asr_ocr/pre.rs: SubtitleSegment 收窄, MIN_SUB_DUR 常量 u32, 循环边界 .into() 适配
- mix_audio/mod.rs: 中间计算保留 i64 宽度, struct 构造处 as u32/i32 收窄
- mix_video/mod.rs: parse_srt_time 改为 u32, json 读取处 as u32
- split_audio: pad_segments 签名改 u32, vad_align 内部保持 u64 计算边界收窄
- translate/tts: JSON parse as_u64() 处统一 as u32

验证: cargo test -p ld-core -p server -p cli 全部通过 (113+ 系列)
bindings: 所有 register_type 导出的 8 类型 ms 字段全部 bigint→number

+++ packages/core/src/cmd/tasks/meta.rs
@@ -183,7 +183,7 @@ fn build_meta_prompt(segs: &[SrtSeg], title: Option<&str>) -> String {
 }
 
 /// 毫秒 → `MM:SS` (如 81200 → 01:21)。
-fn format_ts(ms: u64) -> String {
+fn format_ts(ms: u32) -> String {
     let total = ms / 1000;
     let mm = total / 60;
     let ss = total % 60;
